+++ compiler/src/org.graalvm.compiler.truffle/src/org/graalvm/compiler/truffle/substitutions/TruffleGraphBuilderPlugins.java
@@ -200,6 +200,17 @@ public class TruffleGraphBuilderPlugins {
                 return true;
             }
         });
+        r.register0("inCompilationRoot", new InvocationPlugin() {
+            @Override
+            public boolean apply(GraphBuilderContext b, ResolvedJavaMethod targetMethod, Receiver receiver) {
+                GraphBuilderContext.ExternalInliningContext inliningContext = b.getExternalInliningContext();
+                if (inliningContext != null) {
+                    b.addPush(JavaKind.Boolean, ConstantNode.forBoolean(inliningContext.getInlinedDepth() == 0));
+                    return true;
+                }
+                return false;
+            }
+        });
         r.register0("transferToInterpreter", new InvocationPlugin() {
             @Override
             public boolean apply(GraphBuilderContext b, ResolvedJavaMethod targetMethod, Receiver receiver) {
@@ -242,7 +253,7 @@ public class TruffleGraphBuilderPlugins {
                      * and constant folding could still eliminate the call to bailout(). However, we
                      * also want to stop parsing, since we are sure that we will never need the
                      * graph beyond the bailout point.
-                     *
+                     * 
                      * Therefore, we manually emit the call to bailout, which will be intrinsified
                      * later when intrinsifications can no longer be delayed. The call is followed
                      * by a NeverPartOfCompilationNode, which is a control sink and therefore stops