+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/CompilerDirectives.java
@@ -100,6 +100,18 @@ public final class CompilerDirectives {
     }
 
     /**
+     * Returns a boolean value indicating whether the method is executed in the root of a Truffle
+     * compilation.
+     *
+     * @return {@code false} when executed in the interpreter or in an inlined {@link CallTarget},
+     *         {@code true} when in non-inlined compiled code.
+     * @since 0.28 or earlier
+     */
+    public static boolean inCompilationRoot() {
+        return false;
+    }
+
+    /**
      * Returns a boolean indicating whether or not a given value is seen as constant in optimized
      * code. If this method is called in the interpreter this method will always return
      * <code>true</code>.