+++ truffle/CHANGELOG.md
@@ -9,9 +9,10 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Added `JavaInterop.isJavaObject(Object)` method overload.
 * Deprecated helper methods in `JavaInterop`: `isNull`, `isArray`, `isBoxed`, `unbox`, `getKeyInfo`. [ForeignAccess](http://graalvm.github.io/graal/truffle/javadoc/com/oracle/truffle/api/interop/ForeignAccess.html) already provides equivalent methods: `sendIsNull`, `sendIsArray`, `sendIsBoxed`, `sendUnbox`, `sendKeyInfo`, respectively.
 * Deprecated all String based API in Source and SourceSection and replaced it with CharSequence based APIs. Automated migration with Jackpot rules is available (run `mx jackpot --apply`).
-
+* Added the [inCompilationRoot](http://graalvm.github.io/graal/truffle/javadoc/com/oracle/truffle/api/CompilerDirectives.html#inCompilationRoot--) compiler directive.
 
 ## Version 0.27
+16-Aug-2017
 
 * The Truffle API now depends on the Graal SDK jar to also be on the classpath. 
 * Added an implementation of org.graalvm.polyglot API in Truffle.